+++ src/components/Chatters.tsx
@@ -57,7 +57,15 @@ enum RowType {
 /**
  * Group order used when rendering chatters.
  */
-const GroupOrder: Array<keyof RawChatters> = ['staff', 'global_mods', 'admins', 'moderators', 'vips', 'viewers']
+const GroupOrder: Array<keyof RawChatters> = [
+  'broadcaster',
+  'staff',
+  'global_mods',
+  'admins',
+  'moderators',
+  'vips',
+  'viewers',
+]
 
 /**
  * React State.